Convert PasswordReveal to TypeScript
Closes DEV-1406
